heartbeats: Add support for heartbeats
Host request for heartbeats periodically and EP responds to it. If host
stops receiving heartbeats, it assumes connection to be dead. This
heartbeat mechanism tests the connection bidirectionally. If EP was
periodically sending heartbeats, then it was possible for host packets
to be not reaching EP but only EP's messages reaching host.

Signed-off-by: Saksham Jain <sakjain92@gmail.com>